requirements.txt
# Core scientific stack
numpy<2
setuptools<81
pandas>=2.0,<3.0
matplotlib>=3.7,<3.11
Pillow>=10,<12
PyYAML>=6,<7
opencv-python>=4.8,<5
jupyter>=1.0,<2.0
notebook>=7.0,<8.0
ipywidgets>=8.0,<9.0
tqdm>=4.66,<5.0
seaborn>=0.13,<0.14
pycocotools>=2.0.7,<3.0

# YOLO
ultralytics>=8.2,<9.0

# MMDetection runtime helpers
openmim>=0.3.9,<0.4
mmengine>=0.10,<1.0
mmdet>=3.2,<3.4
